Optional saving comments to file + Multiline comments + Comfy style comments + Comment processing changes
Save comments in png text chunk. Controlled via Settings -> Stable Diffusion -> Stable Diffusion -> Save comments.

/* */ style multiline comment support.

// style single line comment processing (Comfy UI uses these instead of # comments), so I added this for better compatibility.

Added some whitespace trimming regex. This may be bit out of scope for this commit, so commenting it out. Could add a separate option to handle this.

Removed returns in comments.py because the program hangs if saving comments is disabled with these changes active. Not sure why it does this, I'm not knowledgeable enough to get to the bottom of it. These changes force comments to always be processed internally, so #, //, /*, and */ sequences are no longer valid symbols that can be used in prompts.